ML_Driven_Equity_Statistical_Arbitrage_System/requirements.txt
# Core data science and ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Gradient boosting libraries
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep learning libraries (optional)
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Financial data and analysis
yfinance>=0.1.70
pandas-datareader>=0.10.0
ta-lib>=0.4.24
quantlib>=1.26

# Statistical analysis
statsmodels>=0.13.0
arch>=5.0.0

# Alternative data sources
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.8.0

# Parallel processing and optimization
joblib>=1.1.0
numba>=0.56.0
cython>=0.29.0

# Database connections
sqlalchemy>=1.4.0
pymongo>=4.0.0
redis>=4.0.0

# API and web services
flask>=2.1.0
fastapi>=0.75.0
uvicorn>=0.17.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.64.0

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
pytest-xdist>=2.5.0
pytest-benchmark>=3.4.0
coverage>=6.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.17.0

# Jupyter and notebooks
jupyter>=1.0.0
jupyterlab>=3.3.0
ipykernel>=6.9.0
ipywidgets>=7.7.0

# Memory profiling and debugging
memory-profiler>=0.60.0
line-profiler>=3.5.0
py-spy>=0.3.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Serialization and data formats
pickle5>=0.0.11
h5py>=3.6.0
pyarrow>=7.0.0

# Time series specific
pmdarima>=1.8.0
fbprophet>=0.7.1

# MLOps and Model Management
mlflow>=2.0.0
mlflow[extras]>=2.0.0
dvc>=3.0.0
wandb>=0.15.0

# Advanced ML and Deep Learning
torch>=2.0.0
transformers>=4.30.0
optuna>=3.2.0
shap>=0.42.0
lime>=0.2.0.1

# Real-time Data Processing
kafka-python>=2.0.2
redis>=4.5.0
celery>=5.3.0
streamlit>=1.28.0

# Advanced Risk Management
cvxpy>=1.3.0
pyfolio>=0.9.2
empyrical>=0.5.5
riskfolio-lib>=4.3.0

# Alternative Data Sources
alpha-vantage>=2.3.1
quandl>=3.7.0
fredapi>=0.5.0
yfinance>=0.2.0

# Time Series & Forecasting
prophet>=1.1.4
pmdarima>=2.0.0
sktime>=0.21.0
tslearn>=0.6.0

# Visualization & Dashboards
plotly>=5.15.0
dash>=2.12.0
bokeh>=3.2.0
holoviews>=1.17.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
pymongo>=4.4.0
clickhouse-driver>=0.2.6

# Testing & Quality
pytest-benchmark>=4.0.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0

# DevOps & Monitoring
prometheus-client>=0.17.0
grafana-api>=1.0.3
docker>=6.1.0
kubernetes>=27.2.0